risk/get_all_sharpe_symbols/requirements.txt
mysql-connector-python == 8.3.0
requests == 2.31.0